+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/cloud/CloudDownloadMgr.java
@@ -56,7 +56,12 @@ public class CloudDownloadMgr{
   {
     if(isInitialized)
       return;
-    aContext.registerReceiver(completeListener,new IntentFilter(DownloadManager.ACTION_DOWNLOAD_COMPLETE));
+    try {
+      aContext.registerReceiver(completeListener, new IntentFilter(DownloadManager.ACTION_DOWNLOAD_COMPLETE));
+    } catch (IllegalStateException e) {
+      String message = "initialize error: ";
+      Log.e(TAG, message + e);
+    }
     isInitialized = true;
   }
 
@@ -68,7 +73,12 @@ public class CloudDownloadMgr{
   {
     if(!isInitialized)
       return;
-    aContext.unregisterReceiver(completeListener);
+    try {
+      aContext.unregisterReceiver(completeListener);
+    } catch (IllegalStateException e) {
+      String message = "shutdown error: ";
+      Log.e(TAG, message + e);
+    }
     isInitialized = false;
   }
 
@@ -184,8 +194,10 @@ public class CloudDownloadMgr{
                                 ICloudDownloadCallback<ModelType,ResultType> aCallback,
                                 CloudApiTypes.CloudApiParam... params)
   {
-    if(!isInitialized)
-      throw new IllegalStateException("Downloadmanager is not initialize. Call KMManger.initialize before");
+    if(!isInitialized) {
+      Log.w(TAG, "Downloadmanager not initialized. Initializing CloudDownloadMgr.");
+      initialize(aContext);
+    }
 
     synchronized (downloadSetByDownloadIdentifier) {