+++ android/history.md
@@ -1,6 +1,8 @@
 # Keyman for Android Version History
 
 ## 2020-02-14 13.0.6059 beta
+* Bug fix:
+  * Re-initialize CloudDownloadManager when downloading resources after Keyman app is closed (#2635)
 * Change:
   * Restyles the suggestion banner for predictions (#2629)