+++ android/history.md
@@ -1,5 +1,35 @@
 # Keyman for Android Version History
 
+## 2020-02-14 13.0.6059 beta
+* Bug fix:
+  * Re-initialize CloudDownloadManager when downloading resources after Keyman app is closed (#2635)
+* Change:
+  * Restyles the suggestion banner for predictions (#2629)
+
+## 2020-02-13 13.0.6058 beta
+* Change:
+  * Update rest of Android apps' dependency to `androidx.appcompat:appcompat:1.2.0-alpha02` to
+    fix WebView crash on Android 5.0 devices (#2644)
+
+## 2020-02-12 13.0.6057 beta
+* Change:
+  * Update in-app help for 13.0 (#2641)
+  * Update oem dependency to `androidx.appcompat:appcompat:1.2.0-alpha02` to
+    fix WebView crash on Android 5.0 devices (#2640)
+
+## 2020-02-11 13.0.6056 beta
+* No change to Keyman for Android (updated Keyman Web Engine, #2623)
+
+## 2020-02-10 13.0.6055 beta
+* Change:
+  * Update default nrc.en.mtnt model to version 0.1.4 (#2608)
+
+## 2020-02-06 13.0.6054 beta
+* No change to Keyman for Android (updated Keyman Web Engine, #2561)
+
+## 2020-02-04 13.0.6053 beta
+* No change to Keyman for Android (updated Keyman Web Engine, #2559)
+
 ## 2020-01-31 13.0.6051 beta
 * Bug fix:
   * Fix cancelling dictionary update notifications (#2547)